Added possibility to refresh tokens
