+++ server/src/main/kotlin/com/kamelia/jellyfish/plugins/Routing.kt
@@ -1,5 +1,6 @@
 package com.kamelia.jellyfish.plugins
 
+import com.kamelia.jellyfish.rest.auth.authRoutes
 import com.kamelia.jellyfish.rest.user.userRoutes
 import com.kamelia.jellyfish.util.Environment.isDev
 import com.kamelia.jellyfish.util.Environment.isProd
@@ -16,6 +17,7 @@ import io.ktor.server.routing.routing
 fun Application.configureRouting() {
     routing {
         route("/api") {
+            authRoutes()
             userRoutes()
         }
