+++ server/src/main/kotlin/com/kamelia/jellyfish/rest/auth/DTO.kt
@@ -1,0 +1,10 @@
+package com.kamelia.jellyfish.rest.auth
+
+import com.kamelia.jellyfish.rest.core.DTO
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class LoginDTO(
+    val username: String,
+    val password: String,
+) : DTO
