+++ server/src/main/kotlin/com/kamelia/jellyfish/rest/user/DTO.kt
@@ -29,12 +29,6 @@ data class UserDTO(
     val role: UserRole = UserRole.REGULAR
 ) : DTO
 
-@Serializable
-data class UserLoginDTO(
-    val username: String,
-    val password: String,
-) : DTO
-
 /**
  * DTO used to transfer user update details.
  *
