--- server/src/test/kotlin/com/kamelia/jellyfish/ApplicationTest.kt
@@ -1,22 +1,0 @@
-package com.kamelia.jellyfish
-
-import io.ktor.http.*
-import io.ktor.server.testing.*
-import com.kamelia.jellyfish.plugins.configureRouting
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
-class ApplicationTest {
-
-    /*
-    @Test
-    fun testRoot() {
-        withTestApplication({ configureRouting() }) {
-            handleRequest(HttpMethod.Get, "/api").apply {
-                assertEquals(HttpStatusCode.OK, response.status())
-                assertEquals("Hello World!", response.content)
-            }
-        }
-    }
-     */
-}
